Revisando casting para int64 que é nativo bigquery

+++ models/staging/sap_adw/person/stg_sap_adw_address.sql
@@ -18,11 +18,11 @@
 with
     source_address as(
         select
-            cast(addressid as int) as address_id
+            cast(addressid as int64) as address_id
             -- , cast(addressline1 as string) as address_line_1
             -- , cast(addressline2 as string) as address_line_2
             , cast(city as string) as address_city
-            , cast(stateprovinceid as int) as address_state_province_id
+            , cast(stateprovinceid as int64) as address_state_province_id
             -- , cast(postalcode as string) as address_postal_code
             -- , cast(spatiallocation as string) as address_spatial_location
             -- , cast(rowguid as string) as address_row_guid
