+++ models/staging/sap_adw/sales/stg_sap_adw_sales_reason.sql
@@ -18,7 +18,7 @@
 with
     source_salesreason as(
         select
-            cast(salesreasonid as int) as sales_reason_id
+            cast(salesreasonid as int64) as sales_reason_id
             , cast(name as string) as sales_reason_name
             , cast(reasontype as  string) as sales_reason_type
             , date(parse_timestamp('%Y-%m-%d %H:%M:%E*S', modifieddate)) as modified_date
